backend/backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
install==1.3.5
mysqlclient==2.2.4
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
